Native implementation

+++ chvi/wrapper.pyx
@@ -24,17 +24,17 @@ cdef public cpp_vector[size_t] get_state(env):
     return env.state
 
 
-cdef public bool is_terminal(env, cpp_vector[size_t] state):
+cdef public bool is_terminal(env, cpp_vector[double] state):
     return env.is_terminal(np.array(state))
 
 
-cdef public cpp_pair[cpp_vector[size_t],cpp_vector[double]] execute_action(env, cpp_vector[size_t] state, size_t action):
+cdef public cpp_pair[cpp_vector[double],cpp_vector[double]] execute_action(env, cpp_vector[double] state, size_t action):
     assert env.observation_space.contains(state), f"State {state} not part of the observation space"
     assert env.action_space.contains(action), f"Action {action} not part of the action space"
     #env.reset() # not sure if needed
     env.state = env.unwrapped.state = np.array(state) # probably not the most memory-optimized way
     next_state, rewards, _, _ = env.step(action)
-    return cpp_pair[cpp_vector[size_t],cpp_vector[double]] (next_state, np.atleast_1d(rewards))
+    return cpp_pair[cpp_vector[double],cpp_vector[double]] (next_state, np.atleast_1d(rewards))
 
 
 def run(env, discount_factor=1.0, max_iterations=100, epsilon=0.01, verbose=True):
